adrv9009: Add support for Data Offload
This commit updates the testbench to support data offload integration.

The configuration files were updated to emulate the zcu102 use cases.

This commit also addresses the issue in which the the SmartConnect emulating
the CPU will generate an error if it has more than 16 interfaces connected
to it by using an Interconnect instead of SmartConnect.

Signed-off-by: Ionut Podgoreanu <ionut.podgoreanu@analog.com>