i3c_controller: fix stalling on 2024.2
The auto_ack had a limitation of keeping the last value when disable,
that meant it could hold the bus low indefinitely.
Between vivado versions, it would do that at the end of the DAA step.
Create a task to, when disabling, "release" the bus (<= 1'bZ).
Also add one missing set_auto_ack on the private transfer, since the
sub step relies on it.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>