I3C API and fix (#224)
* i3c_controller: Added I3C API and updated testbench

Signed-off-by: Istvan-Zsolt Szekely <istvan.szekely@analog.com>

* i3c_controller: Rename channel to slot on offload I3C API, fix usage

Rename channel to slot to better convey the idea of being slots of a
memory whre the offload CMDs and SDOs are stored.
Fix inversion of write_offload_sdo with write_offload_cmd.
Add a note briefly explaining how does the offload works.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>

* i3c_controller: fix stalling on 2024.2

The auto_ack had a limitation of keeping the last value when disable,
that meant it could hold the bus low indefinitely.
Between vivado versions, it would do that at the end of the DAA step.
Create a task to, when disabling, "release" the bus (<= 1'bZ).
Also add one missing set_auto_ack on the private transfer, since the
sub step relies on it.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>

---------

Signed-off-by: Istvan-Zsolt Szekely <istvan.szekely@analog.com>
Signed-off-by: Jorge Marques <jorge.marques@analog.com>
Co-authored-by: Jorge Marques <jorge.marques@analog.com>