(WIP) Update ad463x TB
All configurations files were updated to include the SPI VIP. New
variables were included to reflect the delays according to the mode
of operation. The TB was also updated to use the SPI VIP API.

TODO: FIFO operations are being inserted into the DMA even though it
is not enabled. Need to find the cause for this bug.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>