testbenches/project/ad463x: Update AD463x FMC TB
PR #1964 updates the TCL scripts for generating
the correct controller for 1-channel or 2-channel
ADC. In that manner, the programmer does not have
to select the useful data when it is using a
single channel ADC.

Updated testbench to support the logic around the
new variable NUM_OF_CHANNEL. Also had to modify
the tcl scripts.

Configuration files are using INTERLEAVE_MODE
instead of NO_REORDER variable. Inserted
ad463x_trigger pin that is controlled by the
GPIO[36] in hardware, here it is set to 0 to
use only PWM signal value.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>